+++ Sources/Paywall/Paywall/Paywall View Controller/WebEventHandler.swift
@@ -234,9 +234,13 @@ final class WebEventHandler: WebEventDelegate {
   }
 
   private func hapticFeedback() {
-    if Paywall.isGameControllerEnabled {
-      return
+    if Paywall.isGameControllerEnabled { return }
+    
+    if #available(iOS 13.0, *) {
+      UIImpactFeedbackGenerator().impactOccurred(intensity: 0.7)
+    } else {
+      UIImpactFeedbackGenerator().impactOccurred()
     }
-    UIImpactFeedbackGenerator().impactOccurred()
+  
   }
 }