+++ Sources/Paywall/Storage/CacheKeys.swift
@@ -9,8 +9,10 @@ import Foundation
 
 enum SearchPathDirectory {
   case cache
-  case documents
+  case userSpecificDocuments
+  case appSpecificDocuments
 }
+
 protocol Storable {
   static var key: String { get }
   static var directory: SearchPathDirectory { get }
@@ -21,7 +23,7 @@ enum AppUserId: Storable {
   static var key: String {
     "store.appUserId"
   }
-  static var directory: SearchPathDirectory = .documents
+  static var directory: SearchPathDirectory = .userSpecificDocuments
   typealias Value = String
 }
 
@@ -29,7 +31,7 @@ enum AliasId: Storable {
   static var key: String {
     "store.aliasId"
   }
-  static var directory: SearchPathDirectory = .documents
+  static var directory: SearchPathDirectory = .userSpecificDocuments
   typealias Value = String
 }
 
@@ -37,7 +39,7 @@ enum DidTrackAppInstall: Storable {
   static var key: String {
     "store.didTrackAppInstall"
   }
-  static var directory: SearchPathDirectory = .documents
+  static var directory: SearchPathDirectory = .appSpecificDocuments
   typealias Value = Bool
 }
 
@@ -45,7 +47,7 @@ enum DidTrackFirstSeen: Storable {
   static var key: String {
     "store.didTrackFirstSeen.v2"
   }
-  static var directory: SearchPathDirectory = .documents
+  static var directory: SearchPathDirectory = .userSpecificDocuments
   typealias Value = Bool
 }
 
@@ -53,7 +55,7 @@ enum UserAttributes: Storable {
   static var key: String {
     "store.userAttributes"
   }
-  static var directory: SearchPathDirectory = .documents
+  static var directory: SearchPathDirectory = .userSpecificDocuments
   typealias Value = [String: Any]
 }
 
@@ -77,6 +79,22 @@ enum Version: Storable {
   static var key: String {
     "store.version"
   }
-  static var directory: SearchPathDirectory = .documents
+  static var directory: SearchPathDirectory = .appSpecificDocuments
   typealias Value = DataStoreVersion
+}
+
+enum LastPaywallView: Storable {
+  static var key: String {
+    "store.lastPaywallView"
+  }
+  static var directory: SearchPathDirectory = .userSpecificDocuments
+  typealias Value = Date
+}
+
+enum TotalPaywallViews: Storable {
+  static var key: String {
+    "store.totalPaywallViews"
+  }
+  static var directory: SearchPathDirectory = .userSpecificDocuments
+  typealias Value = Int
 }